Makefile: Use if statements instead of && and || operators

+++ Makefile
@@ -7,15 +7,14 @@ BUILDDIR = build/
 REMOTEPATH = /var/www/goodmoviefilm.com
 
 $(BUILDDIR): $(SRC)
-	@# `|| true` because make will exit on a failure
-	@[ -d $(BUILDDIR) ] && rm -r $(BUILDDIR) || true
+	@if [ -d "$(BUILDDIR)" ]; then rm -r $(BUILDDIR); fi
 	mkdir $(BUILDDIR)
 	m4 src/index.html.m4 | sed '/^$$/d' > build/index.html
 
 build: $(BUILDDIR)
 
 deploy: build
-	@test $(deploy) || { echo 'Error: deploy must be set.'; exit 1; }
+	@if [ ! "$(deploy)" ]; then echo 'Error: deploy must be set.'; exit 1; fi
 	rsync --rsh='ssh -o StrictHostKeyChecking=no' -r --delete $(BUILDDIR) '$(deploy):$(REMOTEPATH)'
 
 clean:
